+++ xrGame/gamespy/GameSpy_Browser.cpp
@@ -505,7 +505,7 @@ void CGameSpy_Browser::ReadServerInfo(ServerInfo* pServerInfo, void* pServer)
 		ADD_BOOL_INFO(pServerInfo, pServer, *st.translate("mp_si_friendly_names"), G_FRIENDLY_NAMES_KEY);
 
 		ADD_INT_INFO_N(pServerInfo, pServer, 1 / 100.0f, *st.translate("mp_si_friendly_fire"), " %d",
-					   G_FRIENDLY_FIRE_KEY);
+					   (INT)G_FRIENDLY_FIRE_KEY);
 	};
 
 	if (pServerInfo->m_GameType == GAME_ARTEFACTHUNT)
